Finalize dev overlay
