+++ examples/bare/src/app.tsx
@@ -1,9 +1,13 @@
 // @refresh reload
-import { createSignal } from "solid-js";
+import { createSignal, onMount } from "solid-js";
 import "./app.css";
 
 export default function App() {
   const [count, setCount] = createSignal(0);
+
+  onMount(() => {
+    throw new Error('This is an example')
+  })
 
   return (
     <main>
