+++ packages/start/shared/dev-overlay/download.ts
@@ -1,0 +1,6 @@
+export default function download(dataurl: string, filename: string) {
+  const link = document.createElement("a");
+  link.href = dataurl;
+  link.download = filename;
+  link.click();
+}
